downgrade kotlin and fix build
