+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 
-kt = "2.0.20-Beta2"
+kt = "2.0.0"
 kt-serialization = "1.7.1"
 kt-coroutines = "1.8.1"
 
@@ -21,7 +21,7 @@ koin = "3.5.6"
 
 okio = "3.9.0"
 
-ksp = "2.0.20-Beta2-1.0.23"
+ksp = "2.0.0-1.0.23"
 kotlin-poet = "1.18.0"
 
 versions = "0.51.0"
